WebApplication.Feature : last UI and markup fixes

+++ WebApplication/Blocks/TaskList/ViewModels/CardProgressInfoViewModel.cs
@@ -2,8 +2,8 @@
 {
     public class CardProgressInfoViewModel
     {
-        public int CurrentCount { get; set; }
-        public int TotalCount { get; set; }
+        public decimal CurrentCount { get; set; }
+        public decimal TotalCount { get; set; }
 
         public decimal Progress { get { return TotalCount > 0 ? CurrentCount * 100M / TotalCount : 0; } }
     }
